requirements.txt
azure-functions
azure-storage-blob==12.6.0
requests
